requirements.txt
# YouTube 视频下载
yt-dlp>=2024.1.1

# 语音转文本
openai-whisper>=20231117

# 音频处理
ffmpeg-python>=0.2.0

# 数据处理
pandas>=2.0.0

# 进度条显示
tqdm>=4.65.0

# 配置文件处理
pyyaml>=6.0

# 日志处理
loguru>=0.7.0

# 文件操作
pathlib2>=2.3.7
